requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.4.1
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-appconf==1.0.5
django-compressor==4.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
requests==2.28.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
sqlparse==0.4.2
stripe==3.5.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
